etl/requirements/base.txt
pandas==2.2.2
requests==2.31.0
psycopg[binary]==3.2.10
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
pyyaml==6.0.1
